Added slice option
This option allows only sending/receiving a "slice" between vim and the other
process. What's a slice? It's just the lines in the file that contain the
cursor or selection. This is the only part of the file Athame shows to the
user, so doing it this way lets Athame be much more efficient.